Refactored API routes into `user/` and `auth/` subroutes
